Refactor and move operation queue to service
